Fix Stundensatz

+++ src/main/java/com/grash/dto/UserPatchDTO.java
@@ -15,7 +15,7 @@ public class UserPatchDTO {
 
     private String lastName;
 
-    private long rate;
+    private Double rate;
 
     private String phone;
