+++ src/main/java/com/grash/dto/UserResponseDTO.java
@@ -22,7 +22,7 @@ public class UserResponseDTO {
     @ApiModelProperty(position = 3)
     private Role role;
 
-    private long rate;
+    private Double rate;
     private String jobTitle;
 
     private String firstName;
